+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - console: add support for setting aggregation query permissions for ms sql server
 - cli: fix remote schema metadata formatting issues (#7608)
+- cli: fix query collections metadata formatting issues (#7616)
 
 ## v2.5.0-beta.1